+++ test/Gov.t.sol
@@ -7,8 +7,6 @@ import { IManager } from "../src/manager/IManager.sol";
 import { IGovernor } from "../src/governance/governor/IGovernor.sol";
 import { GovernorTypesV1 } from "../src/governance/governor/types/GovernorTypesV1.sol";
 
-import { console2 } from "forge-std/console2.sol";
-
 contract GovTest is NounsBuilderTest, GovernorTypesV1 {
     uint256 internal constant AGAINST = 0;
     uint256 internal constant FOR = 1;
@@ -30,7 +28,7 @@ contract GovTest is NounsBuilderTest, GovernorTypesV1 {
 
     function deployMock() internal override {
         address[] memory wallets = new address[](2);
-        uint8[] memory percents = new uint8[](2);
+        uint256[] memory percents = new uint256[](2);
         uint256[] memory vestingEnd = new uint256[](2);
 
         wallets[0] = founder;
@@ -55,7 +53,7 @@ contract GovTest is NounsBuilderTest, GovernorTypesV1 {
 
     function deployAltMock() internal {
         address[] memory wallets = new address[](2);
-        uint8[] memory percents = new uint8[](2);
+        uint256[] memory percents = new uint256[](2);
         uint256[] memory vestingEnd = new uint256[](2);
 
         wallets[0] = founder;
